backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
langchain-core
langchain-openai
openai
pydantic
python-dotenv